+++ components/manage/MyCompanyTab.jsx
@@ -711,7 +711,7 @@ export default function MyCompanyTab({ filters = {}, onFiltersChange = () => {} 
               <SelectTrigger className="w-20 h-8 text-xs">
                 <SelectValue />
               </SelectTrigger>
-              <SelectContent>
+              <SelectContent position="popper" side="top">
                 <SelectItem value="5">5</SelectItem>
                 <SelectItem value="10">10</SelectItem>
                 <SelectItem value="15">15</SelectItem>
@@ -729,7 +729,7 @@ export default function MyCompanyTab({ filters = {}, onFiltersChange = () => {} 
               <SelectTrigger className="w-20 h-8 text-xs">
                 <SelectValue />
               </SelectTrigger>
-              <SelectContent>
+              <SelectContent position="popper" side="top">
                 <SelectItem value="5">5</SelectItem>
                 <SelectItem value="10">10</SelectItem>
                 <SelectItem value="15">15</SelectItem>